visualization: fix representation of sequence elements
This works, but it is very ugly. Unnecessary data is generated for the
debug message that is then ignored in the DSD Slave. Sequence elements
are no longer pushed on the stack, therefore constructions like
element.parent.parent.get_child(element.activation_reason) are used to
get the sequence element from the action element. This should all be
cleaned up after the RoboCup. I think that a seperate class for the
sequence elements (like AbstractActionElement and
AbstractDecisionElement) should be used instead of the current
situation.